Refactored FirmwareUpgradeConfiguration into its own file
I don't know how many more of these I'm going to do.
